+++ apps/api/src/app/portfolio/calculator/roai/portfolio-calculator-valuable.spec.ts
@@ -120,6 +120,7 @@ describe('PortfolioCalculator', () => {
             averagePrice: new Big('500000'),
             currency: 'USD',
             dataSource: 'MANUAL',
+            dateOfFirstActivity: '2022-01-01',
             dividend: new Big('0'),
             dividendInBaseCurrency: new Big('0'),
             fee: new Big('0'),